Rename net module to accomodate other native libs
